+++ frontend/assets/scss/auth/Connexion.module.scss
@@ -1,0 +1,67 @@
+@import "../core/variables.scss";
+
+.connexion {
+  height: 100vh;
+  display: flex;
+
+  & > * {
+    flex: 1;
+    padding: 20px;
+  }
+
+  aside {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+    background-color: $purple;
+    gap: 30px;
+
+    h1 {
+      font-family: "Wedges", sans-serif;
+      font-size: 6.4rem;
+      color: #ffffff;
+      text-shadow: 2px 2px 0px $pink;
+    }
+  }
+
+  main {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    justify-content: center;
+
+    div {
+      width: 80%;
+    }
+
+    h2 {
+      font-size: 2.4rem;
+      font-weight: 600;
+      color: $purple;
+      margin-bottom: 20px;
+    }
+
+    form {
+      display: flex;
+      flex-direction: column;
+      gap: 20px;
+    }
+
+    span {
+      display: flex;
+      align-items: center;
+      text-align: center;
+      color: #d6d6d6;
+      gap: 12px;
+      margin: 20px 0;
+
+      &::before,
+      &::after {
+        content: "";
+        flex: 1;
+        border-top: 1px solid #d6d6d6;
+      }
+    }
+  }
+}
