+++ frontend/assets/scss/core/core.scss
@@ -1,0 +1,35 @@
+@import "./reset.scss";
+@import "./design_system.scss";
+
+@import url("https://fonts.googleapis.com/css2?family=Inter:ital,opsz,wght@0,14..32,100..900;1,14..32,100..900&display=swap");
+
+@font-face {
+  font-family: "Wedges";
+  src: url("../../fonts/Wedges/Wedges.eot");
+  src: url("../../fonts/Wedges/Wedges.eot?#iefix") format("embedded-opentype"), url("../../fonts/Wedges/Wedges.woff2") format("woff2"), url("../../fonts/Wedges/Wedges.woff") format("woff"),
+    url("../../fonts/Wedges/Wedges.ttf") format("truetype"), url("../../fonts/Wedges/Wedges.svg#Wedges") format("svg");
+  font-weight: normal;
+  font-style: normal;
+  font-display: swap;
+}
+
+html,
+body,
+#__next {
+  height: 100vh;
+}
+
+html {
+  font-size: 62.5%;
+}
+
+body {
+  font-family: "Inter", Arial, Helvetica, sans-serif;
+  font-optical-sizing: auto;
+  font-size: 1.6rem;
+  font-weight: 400;
+  font-style: normal;
+  -webkit-font-smoothing: none;
+  -webkit-font-smoothing: antialiased;
+  -webkit-font-smoothing: subpixel-antialiased;
+}
