+++ frontend/assets/scss/core/variables.scss
@@ -1,0 +1,3 @@
+$purple: #2e1b5c;
+$pink: #fb6ca2;
+$spotify: #1dd760;
