+++ frontend/components/Home.js
@@ -1,45 +1,10 @@
-import { useEffect } from "react";
-import { useRouter } from "next/router";
-import { useDispatch, useSelector } from "react-redux";
-import { setUser } from "../reducers/user.js";
-import spotify from "../modules/spotify";
+import { useSelector } from "react-redux";
 
 function Home() {
-  const router = useRouter();
-  const dispatch = useDispatch();
   const user = useSelector((state) => state.user.user);
-
-  useEffect(() => {
-    if (!router.isReady) return;
-
-    if (router.query.user) {
-      const jsonString = atob(router.query.user);
-      const datas = JSON.parse(jsonString);
-      dispatch(setUser(datas));
-    }
-  }, [router.isReady]);
-
-  useEffect(() => {
-    (async () => {
-      if (user?.email) router.push("/");
-
-      if (user) {
-        // Si le user est connecté faire qqch
-        const datas = await spotify.getMe();
-        console.log(datas);
-      }
-    })();
-  }, [user?.email]);
-
-  const handleLogin = async () => {
-    const response = await fetch("http://127.0.0.1:3000/spotify/login");
-    const datas = await response.json();
-    router.push(datas.redirect_url);
-  };
 
   return (
     <>
-      {!user && <button onClick={() => handleLogin()}>Login</button>}
       <h1>{user && user.email}</h1>
     </>
   );
