+++ frontend/pages/connexion.js
@@ -1,0 +1,7 @@
+import Connexion from "../components/auth/Connexion";
+
+function ConnexionPage() {
+  return <Connexion />;
+}
+
+export default ConnexionPage;
